release-job-migrator: fix release-controller jobName replacement
Due to very simple find replace for updating release-controller job
names, some upgrades would get renamed improperly. For example, in the
updated integration test, the code would've changed
`release-openshift-origin-installer-e2e-gcp-upgrade-4.6-stable-to-4.7`
to
`periodic-ci-openshift-release-master-origin-4.6-e2e-gcp-upgrade-stable-to-4.7`
due to `release-openshift-origin-installer-e2e-gcp-upgrade-4.6` being
migrated and the 4.6-to-4.7 job containing the entire 4.6 job's name in
it. The replacement now does a larger replacement that avoids this
issue.