prowgen: add CPaaS AWS profile
Handle an additional AWS profile where the only difference is AWS
credentials (resulting in using a different AWS account). So the only
difference in jobs should be a different secret mounted to generated
jobs, and different Boskos lease. For all other purposes, the resulting
jobs need to behave as AWS jobs.

This does not attempt to generalize the code enough so that "different"
credentials have first class support. I think that work is related to
Bruno's work on getting rid of cluster profiles entirely, and
generalizing the cluster profile coce would be premature.