ci-operator: resolve imported release image pullspec better
The image reference field in the image import API is documented as being
the correct place from which to pull the imported image, but from
experience we see that it's just the image reference in the stream,
pointing to the image that was imported. As the mechanisms for
authenticating the import are different from the mechanisms used for
using the image directly, it's possible (and likely) that a client that
may import the image may not be able to use it directly. Furthermore,
there's no utility to using the image reference, without a direct pull,
as that is the entire point in doing the import in the first place. By
using out image reference resolution code we can make sure that the
fully resolved pull-spec that we end up using is a local reference and
the kubelet will be able to pull it correctly.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>