steps: Use the pull secret passed to import release from the command
The release extract command needs access to all of the locations for
the images in a release image, not just the release image itself.
The command needs to use the pull secret mounted into the pod by
copying it into the default location and then using login to add
the local cluster's auth.