applyconfig: improve dry runs
Make applyconfig run `oc version` to detect the version of OpenShift we
run against, and select the set of parameters used for dry run based on
the version (it can be overriden by the `--dry-run-method` parameter).

The tool prefers stricter, server-side validation so it is intentionally
selected on various detection errors.

Initially, the detection is opt-in and client dry-run is selected by
default, because our manifests currently do not pass stricter
validation. I want to fix our manifests by running rehearsal with the
auto mode enabled, then we can switch the auto mode to be the default,
without the need to change the actual jobs.