Add azure-arc cluster profile
Add an additional Azure profile for clusters with the Arc-enabled
Kubernetes feature.

For CI purposes, Arc-enabled clusters have absolutely no difference
when compared to other Azure clusters and are created the same way,
however they must run in specific Azure regions where that feature is
already provided.

In this PR we create a new cluster profile to allow us to have a
separate Boskos quota slice only with the supported regions, but
that spins up the cluster as any other Azure cluster in the azure4
quota slice.