applyconfig: handle nonexisting namespaces in server-side dry-run
Previously, applyconfig could spuriously fail when run in server-side
dry-run because such namespaces could be created in an earlier or
current manifest.

This change makes applyconfig analyses the output of oc apply and
searches for messages that indicate that the apply failed because a
namespace was missing.  Then it tries to recover by creating these
missing namespaces and retrying the failed oc apply.  These namespaces
are created annotated for ci-ns-ttl-controller to reap it so unmerged
PRs do not clutter the cluster. If a manifest with the namespace is
actually applied (post-merge), the annotations are cleaned up.