prowgen: explicitly opt out of pod-utils with artifact dir
There's one major difference between the current artifact gathering code
and using the Prow pod utilities: the former allows for users to
configure their artifact directory, whereas the latter expects to be
declarative itself and for the user to simply consume the directory as
an environment variable. As we work to make the latter approach default,
we first need to opt-out if we notice a user is setting their directory
to something non-standard.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>