pkg/registry: merge test/workflow list/dict fields
While technically correct according to the [specification][0],
completely overriding the workflow fields is sometimes not what is
desired when fields are added to the test in the `ci-operator`
configuration file.  [DPTP-1824][1] documents one of these cases: when
an `env` section is added to a test, the intent is likely to augment the
workflow section, but it is instead completely overridden.

The behavior of configuration resolution is changed so that the two
sections are now merged for list and dictionary fields, namely:

- `env`/`TestEnvironment`: parameters in the test override those in the
  workflow if they target the same environment variable.  Otherwise,
  the resulting parameter list is the combination of both sections.  An
  existing value can be cleared by setting `default` to the empty
  string.  It is not possible to revert the effects of the workflow
  section (i.e. use the original default value), but that is unlikely to
  be useful.
- `dependencies`/`TestDependencies`: same treatment as the previous item.
- `leases`/`[]StepLease`: resources in the test must not target an
  environment variable already present in the workflow.  Otherwise, the
  resulting lease list is the combination of both sections.

[0]: https://docs.ci.openshift.org/docs/architecture/step-registry/#ci-operator-test-configuration
[1]: https://issues.redhat.com/browse/DPTP-1824