secret-wrapper: handle writable home
There are a number of changes we make to the environment of the process
under test. Some occur in the `secret-wrapper` tool, while some form an
ugly Bash preable. As the content of these changes continues to grow and
become more complex, we should aim to make all of them in the Go tool.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>