cmd/testgrid-config-generator: All release periodics should appear in testgrid
Release periodics are all informing, and testgrid visibility is a
key part of TRT automation. Jobs that match the standard pattern
should appear in testgrid unless there is a reason.

We recently standardized naming for all newly generated periodic
jobs to have a predictable structure. We use 'ci' and 'nightly' and
'okd' to identify the job stream in the new structure, so ensure the
generator uses those names to assign the right dashboard type (ci
and nightly both show up in ocp dashboards).

Upgrade jobs should be part of the release they define, not generic
informing. The new naming structure puts the target release (newer)
in the first position, which allows the generator to correctly pick
the right target.

Add jobs without a version to the generic informing dashboard, as
originally intended.

This commit allows the allow list to simply become an override of
which dashboard a small set of jobs should appear in. After this
change the allow list should be a very, very small set.