ci-operator: stop allowing users to set artifact dir
No longer are users able to set the explicit directory for:
 - `ci-operator` outputs
 - simple Pod tests
 - multi-stage test steps

`ci-operator` itself will read the `${ARTIFACTS}` value to determine
where to put its own artifacts (like the output jUnit files). This is in
effect what already happens, as this is the fallback to when the flag is
not passed, and we do not pass it in generated jobs. The flag will
remain in a deprecated manner so as to not break users.

All test steps *except* for `Template` tests are already using the
test-infra pod-utils to upload, and therefore are not allowed to set a
non-standard directory, anyway. Nothing changes for these steps. It is
now not possible to opt out of this by telling `ci-operator` not to use
pod-utils.

We continue to use the deprecated artifact fetching sidecar code for
Templates, but only for templates.

This will cause a breaking change to prowgen as our previous automation
explicitly wrote our defaulted value into all configs even if they never
set it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>