ci-operator: censor all secret data in sidecar
The Prow pod utilities can now censor secret data before uploading
anything to the cloud, both in plain-text files and in archives that the
user creates during a test run. We know that the set of sensitive data
available to test Pods for a job is limited to the set of Secrets in the
Namespace that ci-operator created for the job, so by mounting them all
into the test Pods we can effectively guard against all secret leaks in
the future.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>