ci-operator: always promote the build cache
Lots of valuable work is done when binaries are built from the source
code in a test. We can promote the resulting image to a build cache
repository and use it as the `build_root` for subsequent jobs (if the
underlying `build_root` that we originate from has not changed) and
thereby save enormous amounts of time in both incremental cloning and
incremental compilation.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>