multi_stage: write commands.sh to configmap
Instead of running commands directly with `bash` when running multistage
tests, `ci-operator` should instead execute the script directly so that
the shebang specifying what shell to use is followed (it is ignored when
run directly by the shell) and to remove the multistage test steps'
dependance on `bash`, which does not exist in certain images, usually
alpine based.

This PR now runs commands by creating a configmap containing the
commands script for each step in a multistage test and mounting it to a
known location in the pod. Then, the script for the current step is
configured as the entrypoint for the step's pod.