Add vault support to secret bootstrapper
This is very straightforward except for the getUnusedBWItems check. That
check verifies that all of fields, attachments and password are used, if
present. As Vault just has a k-v map and doesn't know to what bitwarden
type a field maps we do the following:
* We store all keys vault has
* For all of field/attachment/password in the config file, we ask the
  secret provider if it has them
* This will make the vault provider mark those fields as in-use
* At the end, we add another check, "superfluous fields". The Vault
  implementation will use that to report all fields it didn't get asked
  about

This allows us to re-use all tests and thus be confident that it works
for Vault as well. The error message however is slightly different for
Vault.