CI Operator: Use single-pod watches instead of polling
This change makes the client used by the ci-operator transparently
create a single-object watch for all Pods requested through it and then
serve that request for the pod. The mechanic allows to trivially add
other object types there.

This alows us to keep using a single, generic client everywhere without
needing to resort to polling.