types: empty string is not a bundle image
This commit improves checking of the `IsPipelineImage` function by
identifying all index images instead of just unnamed ones and preventing
`IsBundleImage` from returning `true` for empty image names, which was
causing broken dependencies when the PR containing the previous commit
in this PR merged. Further, this commit prevents the MultiStage step's
run function from adding a dependency for the `from` image if `from` is
empty (which is why `IsBundleImage` was given an empty `imageName`
string in the first place).