ci-operator: label digest of image we built on top of
In the future, `ci-operator` will optionally re-use a `bin` image to get
access to the clone and build cache that is stored there after the image
is built. In order to be able to tell that a cached image is
out-of-date, we need to label the version of the root it's built on top
of and check that against the version of the root we resolved at
runtime.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>