ci-operator: clean up build root step generation
There's no place where any of the weird "test-base" stuff is used and
the previous commit validates as much. I'm not sure what it was for and
it's not documented anywhere. I removed that handler and fixed the git
base / project build options to pass through the full struct, instead of
cherry-picking some options and silently dropping the rest.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>