ci-operator: optionally use build cache for the root
When the user asks us to, we can determine if the build cache we have
published for the repo is up-to-date. If it is, we should use it for our
build root instead of the build root that would otherwise be used. In
cases where we can't resolve it or the cache is stale, fall back to the
original root.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>