ci-operator: initialize a more intelligent logger
Default censoring that is provided by `sidecar` has one draw-back: the
set of secrets that must be censored is loaded at start-up and must be
known ahead of time in order to be provided to the process. Inside of
the `ci-operator`, we routinely touch secret values via the Kubernetes
API at runtime. This means that the total set of secret values we may
possibly leak grows over time. Due to this, we cannot solely rely on the
censoring implemenation that runs out-of-process.

Also, today the `ci-operator` uses the `log` library for logging, in
contrast to pretty must every other part of the codebase and larger
ecosystem (Prow, Boskos, etc...). Censoring for the `logrus` library is
implemented, tried, and tested in production for a long time. We are
better off by migrating to `logrus` over time in order to consolidate
our codebase.

We have a similar issue in the `ci-secret-bootstrap` tool and already
have the `DynamicCensorer` utlity to handle this case. This patch does
two things:

- create an initial implementation of a `logrus` logger for
  `ci-operator` that not only censors by default but also shows only the
  most important messages to users in stdout, while at the same time
  funneling more specifc/spammy messages to a file which an
  administrator can read after a job failed.
- wire through a client shim that records the values of Secrets that
  `ci-operator` touches in the Kuberntes API while it runs, in order to
  ensure that the censoring provided by the logger cannot leak anything
  sensitive

Follow-up PRs will begin to move usage of `log` to `logrus` in our
code-base.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>