ci-operator: stop censoring SA secrets
It turns out that these secrets contain, among other things, the
namespace for the SA. This ends up used in a multitude of places,
including the cluster names we start, etc. Users do not have the RBAC to
look at these secrets themselves, and they would have to guess the
random hash in the name of the secret to mount it using `credentials`,
so there is effectively no real way for a user to access the value of
the SA credentials in their namespace. If anything leaked here, it would
also not be useful past the TTL of the namespace. We can therefore omit
these mounts from censoring.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>