CI-Secret-Bootstrap: Sync user secrets from Vault
This change adds support to the secret bootstrapper to sync user secrets
from Vault. It will use specific keys in the Vault secret to figure out
the namespace/name it should sync to. The clusters need to be explicitly
configured. It is allowed to have secrets that contain both data from
the secret bootstrap config (owned by dptp) and user secrets but dptp
data takes precedence and we will error if an attempt to overwrite it is
made.