ci-operator: use logrus in main.go
This change does a couple of things:

 - use logrus instead of log
 - change log messages like "warning: ..." into first-class logrus
   Warn() calls
 - capitalize first letters, add periods as needed
 - change to use WithError() instead of printing it into the message

Notably, I translate `log.Print()` to `logrus.Info()`, so there should
be no change to the lines output to the user-facing log. In my opinion,
some of these meessages are spammy and not useful for a user, but we can
change that in the future.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>