ci-operator: add a utility to write censored artifacts
We want to make sure that we censor anything we're writing to artifacts
from this process as we often interact with secrets during runtime for
the first time, so we can't be certain that out-of-process censoring
will do anything sensible. A helper makes it easier for an engineer to
do the right thing.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>