rehearse: deduplicate the list of steps that trigger rehearsals
Previously the list of steps, each of which triggered a rehearsal, could
contain duplicates: this happened when two steps were changed and they
shared an ancestor. Such ancestor was then added twice and thus
triggered two rehearsals. Deduplicate the list of affected steps, and
extract the whole "discover the set of affected step" into a separate
method.