entrypoint-wrapper: label secret to be ignored in censor
We do not read/update/write the shared directory secret in the
entrypoint wrapper, we just create it from the values on disk since that
is technically authoritative for the data. It causes us to lose the
labels, though, so we need to set the censoring skip label here as well
or we will end up censoring random data that's useful to users.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>