ci-operator: update the label set we use on Pods
The intent of adding labels to the Pods we create was always two-fold:
first, help administrators identify Pods as they relate to jobs and
steps. Second, allow for the labels to end up in Prometheus to help us
track metrics across job runs. Unfortunately, the previous label set
fell short in the second case:

 - we added the full job name as a label, which was trimmed to fit in
   most cases and in some cases the trimming left only the org/repo/etc
   parts, making it not actually useful as an identifier
 - we did not label the target that ci-operator executed, just fields
   like the full job name which might include it as a suffix
 - we did not label the multi-stage step name, now allowing us to
   discern individual usage metrics
 - we labeled the build and Prow job IDs, which are unique and not
   ultimately useful to selecting anything

This patch changes the label set slightly to make it more useful.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>