pod-scaler: a server for automated resource requests
This commit introduces a new server to automate setting resource
requests for our batch workloads by looking at the resource consumption
of workloads in the past for the same identifiers.

This server is split into two modes - a producer that periodically loads
data from Prometheus, digests it and pushes it into the cloud, and a
consumer which loads the data from the cloud, digests it into resource
recommendations and can also serve a web UI. The producer has previously
been merged, this patch provides the mutating admission webhook server
to set resource requests and the web UI server.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>