pod-scaler: a server for automated resource requests
This commit introduces a new server to automate setting resource
requests for our batch workloads by looking at the resource consumption
of workloads in the past for the same identifiers.

This server is split into two modes - a producer that periodically loads
data from Prometheus, digests it and pushes it into the cloud, and a
consumer which loads the data from the cloud, digests it into resource
recommendations and can also serve a web UI. The producer has previously
been merged, this patch provides the mutating admission webhook server
to set resource requests. The UI server for viewing the data will come
next.

While we roll out this feature, we need to be doing it in a
non-disruptive way. So, this patch will set requests and limits if and
only if the values we compute are larger than those already set on the
Pod by a user. We will then remove the configuration from the release
YAMLs, to allow the server to do what it needs to, always allowing users
to add configuration to explicitly turn up the requests past what we
think is appropriate.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>