pod-scaler: a server for automated resource requests
This commit introduces a new server that will automate setting resource
requests for our batch workloads by looking at the resource consumption
of workloads in the past for the same identifiers.

This server is split into two modes - a producer that periodically loads
data from Prometheus, digests it and pushes it into the cloud, and a
consumer which loads the data from the cloud, digests it into resource
recommendations and can also serve a web UI.

This commit contains the producer code. In order to provide an estimate
of resource usage for containers in a CI job, this server analyzes
metrics from previous executions of similar containers. Aggregate
statistics are used to provide resource request recommendations by
digesting prior metrics. It is assumed that, for a sufficiently similar
container, resource usage will not vary much across executions - we
expect this to be true for e.g. all executions of unit tests for some
branch on a repository. This assumption allows for samples from all
executions to be treated as one dataset with a single underlying
distribution, so that aggregation can be done on the larger dataset to
yield higher-fidelity signal. The overall size of the raw data, however,
quickly grows unmanageable. In order to operate efficiently on this
dataset we therefore store compressed histograms for each execution
trace. This allows us to reduce the data footprint while continuing
to allow for dataset merging and aggregation. The Circonus log-linear
histogram is used as it's performant, accurate, efficient and open-
source.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>