prowgen: add release-informing job type
This PR adds support for defining jobs as release-controller jobs to
prevent confusion when a periodic job is configured via
release-controller instead of a prowjob or ci-operator config. A
`release-controller` PR will be created to validate that all jobs
defined in the release-controller periodic and verification jobs are
appropriately labelled as release-controller jobs and that all jobs with
the label are in the release-controller configs.