ci-operator: disable caching on empty input SHAs
Prowjobs with `extra_refs` do not contain SHAs (periodics and
rehearsals, among others), so we end up using an empty string as a hash
input and sometimes hit cache when we should not.

We can use time as an artificial input in such case, effectively
disabling the cache.

Alternatively, we could fix up Prowjob producers to always resolve SHAs
for refs, but even in that case this chance protects us from silently
doing the wrong thing on empty input.