Vault accessory: Use testing.T instead of testhelper.T
testhelpers.T must be run from the testframeworks Run method to ensure
proper synchronization, otherwise it will not wait correctly and swallow
Errorf and Fatalf calls, making failing tests appear as if they
succeeded.

This changes the accessory to also allow testing.T, and adds a strongly
worded warning when using testhelper.T. The main drawback is that
failures in the accessory can not Fatalf the test anymore, as that is
not allowed from a distinct goroutine. In all practical terms this is of
little relevance for tests that have a runtime of <10 seconds like the
Vault tests.