pod-scaler: store data for three types of Pods
In the end it turns out to be useful to consider three classes of Pods
on our clusters:

 - Pods started by Prow/plank to run ProwJobs
 - Pods started by ci-operator that do not relate to a specific step
 - Pods started by ci-operator that implement a multi-stage step

Different sets of labels are used to identify these classes of Pods.
Different assumptions may be mafe about them, for instance:

 - Pod name is not relevant for Pods started by Prow
 - Pods started by ci-operator that aren't for steps may be sharded by
   the target, or may be equivalent across a ci-operator configuration

This patch changes the GCS storage mechanism to record memory and CPU
usage for each of the three types. Secondary indices (like step-based
ones) are removed from the storage type, they can be efficiently and
quickly computed at run-time by consumers of the data.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>