ci-operator: label release pods with their release name
Release pods vary only by the CI Operator configuration, not the
specific target that ends up running them. It is important to be able to
detect that a Pod is being run for a release, in some manner that's less
fragile than searching for a prefix on the Pod name.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>