[DPTP-2119] steps: clean up cluster claims when test ends (#1905)
* [DPTP-2119] steps: clean up cluster claims when test ends

This PR adds deletion functionality for cluster claims to be deleted
when a test using one ends. This is handled in the same way as leases,
where the test step is wrapped in the cluster claim step. This allows
the cluster claim step to delete the claim after the test ends.

* chanhes based on review

* changes based on review