Fix operator e2e tests and add additional validation around lease validation to skip.
With DPTP-2012 we want to add validations that, when a test includes a trap command, a grace_period is also set, and that when best_effort is true, a timeout is also set.Remove funky var.

Simplify if statement.

Run make generate.

Aggregate all validation errors.
Update regex to match start-of-line or whitespace.

With DPTP-2012 we want to add validations that, when a test includes a trap command, a grace_period is also set, and that when best_effort is true, a timeout is also set.

With DPTP-2012 we want to add validations that, when a test includes a trap command, a grace_period is also set, and that when best_effort is true, a timeout is also set.