pkg/results: flatten error trees
The `Aggregate` error type from `apimachinery` explicitly does not
support `Errors.As`:

```go
// Errors.As() is not supported, because the caller presumably cares
// about a specific error of potentially multiple that match the given
// type.
```

This breaks the causal chain in the `FullReason` method.  This change
adds a function that flattens trees of arbitrarily-nested
`results.Error` and `Aggregate`-like errors into a list while preserving
the hierarchy, so that all errors reported have the correct reason.