Bootstrap item validation: Correctly strip prefix
We insert the VaultDPTPPrefix into all item names during deserialization
in the bootstrapper and give the generator the same prefix as CLI arg.

This means that the generator config has the items without that prefix
so we need to strip it before we check in the generator config if it
exists.