DPTP-2225: do not log file not found errors for resolver
Due to a potential race condition between the configresolver reloading
configs or the registry and the k8s configmap updater removing old
symlinks for file that were removed/moved, we occassionally see file not
found errors during reload. These are harmless and we should ignore them
instead of returning an error.