CI-Operator: Create imports for root image if FromRepository is enabled
This change makes the CI-Operator create testimagestreamtag imports for
the root image if the FromRepository setting is set. This is best-efford
only and will log warnings on errors but not fail.

It is also insufficient by itself, we need to deploy the CRD, add RBAC
and extend the test-image-distribution-controller to watch
testimagestreamtagimports in all clusters.