Determinize Prow config: Fix dir for additional plugin configs
Before this change, we looked at`filepath.Dir(config.PluginConfigFile)`
for additional plugin configs. Since `config.PluginConfigFile`` is just a
filename, this resolves to `.` which is wrong (but happens to include
the correct files, as they are in a subdir) and always fails in Docker
with an error like `failed to walk .: lstat proc/1/fd/6: no such file or directory"`.

This changes it to correctly look at the dir of the entire plugin config
file path instead, which fixes above error.