Test image distributor: Reconcile all imagestreamtags from new config
This makes the test image distributor subscribe to config changes and
whenever a new config was added, it will reconcile all images in it.

Before this change we would only allowlist the images on config change
but not actually distribute them until they change (or the controller
restarts). After this change, we will immediately distribute them.